Add main.go and HTTP server setup
This commit chiefly includes the addition of `main.go` which sets up an HTTP server using `echo` framework. It also creates an HTML rendering function, sets up a single route for a counter, and initiates the server on port `42069`. Additional changes contain the introduction of `.gitignore` entries for `.idea`, and setting up `go.mod` and `go.sum` files for dependency management. Lastly, an `index.html` file is created in the `views` directory.